+++ pom.xml
@@ -48,6 +48,12 @@
 	    <artifactId>postgresql</artifactId>
 	    <version>42.2.23</version>
 	</dependency>
+	<dependency>
+  		<groupId>javax.servlet</groupId>
+  		<artifactId>javax.servlet-api</artifactId>
+  		<version>3.1.0</version>
+  		<scope>provided</scope>
+	</dependency>
   </dependencies>
 
   <build>
