+++ src/main/webapp/elementos/footer.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 	<body>
-	    <footer class="fixed-bottom footer mt-auto py-3 bg-light">
+	    <footer class="footer mt-auto py-3 bg-light">
 	      <div class="container">
 	        <span class="text-muted">2021 ArtDigital</span>
 	      </div>
