+++ src/main/webapp/sobre.jsp
@@ -1,0 +1,111 @@
+<%@ page language="java" contentType="text/html; charset=ISO-8859-1"
+	pageEncoding="ISO-8859-1"%>
+<!DOCTYPE html>
+<html>
+<head>
+
+<meta charset="UTF-8">
+<meta http-equiv="X-UA-Compatible" content="IE=edge">
+<meta name="viewport" content="width=device-width, initial-scale=1.0">
+
+<%@include file='elementos/bootstrap5.html'%>
+<title>Sobre | ArtDigital</title>
+
+</head>
+<body>
+	<%@include file='elementos/header.jsp'%>
+
+	<div class="container">
+		<div class="row">
+			<p class="fs-1 fw-bold">ArtDigital</p>
+			<p class="fs-4">A ArtDigital foi desenvolvidada voltada para a
+				área do artesanato, com intuito de proporcionar comodidade e novas
+				oportunidades para artistas regionais, disponibilizando uma
+				plataforma onde os artesãos irão divulgar suas obras de artes que
+				poderão ser vendidas em todo o território nacional.</p>
+			<p class="fs-3 fw-bold">Qual o principal problema nós resolvemos
+				?</p>
+			<p class="fs-4">Uma quantidade massiva de artesãos produz suas
+				obras e enfrentam dificuldade em divulgá-las para venda. A ideia do
+				projeto é criar uma plataforma que faça o elo entre apreciadores da
+				arte, possíveis compradores, e os artesãos, fornecedores. A
+				interação com a ferramenta será simples e de rápido aprendizado.</p>
+			<p class="fs-3 fw-bold">Qual é a nossa proposta ?</p>
+			<p class="fs-4">A proposta do nosso Aplicativo Web é gerar
+				visibilidade para os Artesões. Visibilidade essa que talvez ele
+				mesmo não saiba como alcançar. O intuito é torna-los importantes
+				diante daquilo que eles são, cultura. Visamos atingir todas as
+				idades, sem distinção alguma, porque talento não tem idade. O
+				ArtDigital vem para mudar a vida dos comeciantes de artes, traze-los
+				para a nova era, a era da informação. Atualmente, o aplicativo tem
+				um ponto muito diferente das outras plataformas, dar opções de como
+				por um preço em sua arte, que muitas vezes o própria artista não
+				sabe como se faz, mesmo com anos de experiência. Novas formas de
+				divulgar e novos jeitos de secomunicar com o comprador.</p>
+			<p class="fs-3 fw-bold">Como funciona o nosso plano de assinatura
+				?</p>
+			<p class="fs-4">Será cobrado uma taxa de comissão de venda sobre
+				o valor do pedido total, percentual de 10% nas compras que forem
+				realizadas pelo Artdigital Pay e o impulso de produtos que permite
+				melhorar a classificação de busca dos produtos, aumentando o número
+				de visualizações e venda em potencial. Será cobrado valores
+				estipulados em tabela, podem impulsionar entre 7, 15 e 30 dias com
+				os seguintes valores: R$ 9,99 R$ 19,99 R$ 29,99, respectivamente.</p>
+				<p id="helpId" class="fs-3 fw-bold">Como funciona a ajuda na hora de precificar os produtos?</p>
+			<p class="fs-4">
+				Primeiro a pessoa decide o quanto quer ganhar de salário e divide isso pelas horas que trabalha por mês.
+				<br>
+				Exemplo:
+				<br>
+				Um vendedor quer ganahr R$ 1008,00 de salário e quer trabalhar 06:00 horas diárias, de segunda à sexta.
+				<br>
+				Ele preenche os campos de: Dias de trabalho, horas trabalhadas e valor da hora. Segue o cálculo:
+				<br>
+				<strong>Dias de trabalho</strong> => 24 dias
+				<br>
+				<strong>Horas trabalhadas</strong> => 06:00 horas
+				<br>
+				<strong>Valor da hora</strong> => R$ 7,00
+				<br>
+				O valor da hora é R$ 7,00 porque ele deseja ganhar <strong>R$ 1008,00</strong> por mês.
+				<br>
+				Cálculo: (24 x 6) x 7 = 1008
+				<br>
+				Depois de saber o valor da hora, precisa adicionar os custos.
+				<br>
+				Para produzir um biquini de crochê precisa-se de 2 novelos de lã, cada um vai sair por R$ 18,00. Isso pode variar de acordo com a marca e o tipo de artesão que vai usar.
+				<br>
+				Além disso, precisa de tecido para fazer o forro, uma embalagem, cola e linha de costura.
+				<br>
+				Material:                  							                                                                                
+  				<br>
+  				<strong>R$ 18,00</strong> - Dois Novelos;							
+                <br>
+                <strong>R$ 10,00</strong> - Uma Cola;             					         						    
+                <br>
+                <strong>R$ 2,50</strong>  -  Meio metro de Tecido para o forro;     
+  				<br>
+  				<strong>R$ 1,50</strong>  -  Embalagem;  							
+                <br>
+                <strong>R$ 2,00</strong>  -  Linha de costura.                                   					
+  				<br>       
+  				Total = <strong>R$ 40,00</strong>  
+  				<br>  
+  				Tempo:
+  				<br>
+  				Horas trabalhadas (Horas) = <strong>07:00 horas</strong>
+  				<br>         
+  				Valor da hora (R$) = <strong>R$ 7,00</strong>      
+  				<br> 
+  				7 x 7 = <strong>R$ 49,00</strong>               
+  					<br> 
+  					R$ 49 + R$ 40 = R$ 89,00    
+  					<br> 
+  					O valor do biquíni vai ser no valor de: <strong>R$ 89,00</strong>.                                               
+			</p>
+		</div>
+	</div>
+
+	<%@include file='elementos/footer.html'%>
+</body>
+</html>
