--- src/test/java/utils/HibernateTestUtil.java
@@ -1,47 +1,0 @@
-package utils;
-
-import java.util.Properties;
-
-import org.hibernate.SessionFactory;
-import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
-import org.hibernate.cfg.Configuration;
-import org.hibernate.cfg.Environment;
-
-import entity.Categoria;
-import entity.Produto;
-import entity.Usuario;
-
-public class HibernateTestUtil {
-    private static SessionFactory sessionFactory;
-    
-    public static SessionFactory getSessionFactory() {
-        if (sessionFactory == null) {
-            try {
-                var configuration = new Configuration();
-                
-                var settings = new Properties();
-                settings.put(Environment.DRIVER, "org.h2.Driver");
-                settings.put(Environment.URL, "jdbc:h2:mem:test");
-                settings.put(Environment.USER, "sa");
-                settings.put(Environment.PASS, "");
-                settings.put(Environment.DIALECT, "org.hibernate.dialect.H2Dialect");
-                settings.put(Environment.SHOW_SQL, "true");
-                settings.put(Environment.CURRENT_SESSION_CONTEXT_CLASS, "thread");
-                settings.put(Environment.HBM2DDL_AUTO, "create-drop");
-                
-                configuration.setProperties(settings);
-                configuration.addAnnotatedClass(Usuario.class);
-                configuration.addAnnotatedClass(Produto.class);
-                configuration.addAnnotatedClass(Categoria.class);
-
-                var serviceRegistry = new StandardServiceRegistryBuilder()
-                    .applySettings(configuration.getProperties()).build();
-
-                sessionFactory = configuration.buildSessionFactory(serviceRegistry);
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-        return sessionFactory;
-    }
-}
